refactor(dashboard): reorganize imports and update component references
- Update list export to use the correct page reference
- Reorganize imports in CheckedTotalPieChart, ItemVariationRow, RecentListSquareCard, and ItemPriceComparisonPage components for consistency
- Change formatCurrency import path to a shared utility
