+++ src/app/(authenticated)/list.tsx
@@ -1,1 +1,1 @@
-export { default } from '@/features/list-items/page';
+export { default } from '@/features/list/page';
