+++ src/features/dashboard/components/recent-list-square-card.tsx
@@ -1,12 +1,11 @@
-import { Pressable, View } from 'react-native';
-import { IconShoppingCart } from '@tabler/icons-react-native';
-
 import { Icon } from '@/components/ui/icon';
 import { Text } from '@/components/ui/text';
-import { formatCurrency } from '@/features/list-items/utils/formatters';
-import { iconMap } from '@/features/lists/utils/icon-map';
 import { getAccentColorCardClasses } from '@/features/lists/utils/accent-colors';
+import { iconMap } from '@/features/lists/utils/icon-map';
 import { cn } from '@/lib/utils';
+import { formatCurrency } from '@/utils/formatters';
+import { IconShoppingCart } from '@tabler/icons-react-native';
+import { Pressable, View } from 'react-native';
 
 import type { DashboardRecentListCard } from '../types';
